fix: Carbon createFromFormat trailing data + hospital_prescriptions FK type mismatch
- HomeController line 286: prepend day-of-month so Carbon has unambiguous date input
- patient_id must match contacts.id type (unsignedInteger, not unsignedBigInteger)

+++ database/migrations/2026_04_19_000001_create_hospital_prescriptions_table.php
@@ -15,7 +15,7 @@ class CreateHospitalPrescriptionsTable extends Migration
         Schema::create('hospital_prescriptions', function (Blueprint $table) {
             $table->increments('id');
             $table->unsignedInteger('business_id');
-            $table->unsignedBigInteger('patient_id');           // contacts.id
+            $table->unsignedInteger('patient_id');               // contacts.id
             $table->unsignedBigInteger('doctor_id')->nullable(); // users.id
             $table->unsignedBigInteger('consultation_id')->nullable();
             $table->unsignedInteger('variation_id')->nullable(); // variations.id
